refactor: move components to commons folder
