+++ src/components/app/pagePetitionDetails/common/debugger/petitionDataEditor.tsx
