+++ src/components/app/pagePetitionDetails/common/summary/mobileGradient.tsx
@@ -1,0 +1,5 @@
+export function PetitionMobileGradient() {
+  return (
+    <div className="absolute -bottom-10 left-0 right-0 -z-10 h-12 bg-gradient-to-b from-background to-transparent"></div>
+  )
+}
