+++ src/components/app/pagePetitionDetails/common/wrapper.tsx
@@ -1,0 +1,3 @@
+export function PetitionDetailsWrapper({ children }: { children: React.ReactNode }) {
+  return <div className="standard-spacing-from-navbar container mx-auto max-w-6xl">{children}</div>
+}
